Fix running "rails new --css bootstrap" on 7.1
Rails 7.1 included a [change][1] to allow using importmaps along with
all cssbundling options. However, the Bootstrap installer was never
updated to take this new default into effect (and is currently broken
because of this).

This commit adds the additional configuration required to use the
Bootstrap npm package with importmaps so that "rails new" generates an
application that runs without errors.

[1]: rails/rails@84458a8577e09621205f2ea86f570745033105f5