Use Thor's `apply` instead of a prerequisite task (#150)
The `css:install:shared` task serves only as a prerequisite for the
other installer tasks; it should not be run on its own (nor listed with
`rake --tasks`).  By replacing this task with corresponding calls to
Thor's `apply` method, we avoid the overhead of running
`bin/rails app:template` (and `bundle install`) multiple times.